+++ crystal/__init__.py
@@ -1,5 +1,5 @@
 __version__ = "1.0.0"
-__build__ = "016"
-__release__ = "Crystal V1.00 Build 016"
+__build__ = "017"
+__release__ = "Crystal V1.00 Build 017"
 
 __all__ = ["__version__", "__build__", "__release__"]